feat: update cache on message added
