+++ client/src/components/ChatRoomScreen/index.tsx
@@ -1,4 +1,3 @@
-import { defaultDataIdFromObject } from 'apollo-cache-inmemory';
 import gql from 'graphql-tag';
 import React, { useCallback } from 'react';
 import styled from 'styled-components';
@@ -6,9 +5,9 @@ import ChatNavbar from './ChatNavbar';
 import MessageInput from './MessageInput';
 import MessagesList from './MessagesList';
 import { History } from 'history';
-import { ChatsQuery, useGetChatQuery, useAddMessageMutation } from '../../graphql/types';
-import * as queries from '../../graphql/queries';
+import { useGetChatQuery, useAddMessageMutation } from '../../graphql/types';
 import * as fragments from '../../graphql/fragments';
+import { writeMessage } from '../../services/cache.service';
 
 const Container = styled.div`
   background: url(/assets/chat-background.jpg);
@@ -42,10 +41,6 @@ interface ChatRoomScreenParams {
   history: History;
 }
 
-interface ChatsResult {
-  chats: any[];
-}
-
 const ChatRoomScreen: React.FC<ChatRoomScreenParams> = ({ history, chatId }) => {
   const {
     data, loading
@@ -63,7 +58,6 @@ const ChatRoomScreen: React.FC<ChatRoomScreenParams> = ({ history, chatId }) => 
 
       // tslint:disable-next-line:no-shadowed-variable
       const chat = data.chat;
-
       if (chat === null) { return null; }
 
       addMessage({
@@ -76,82 +70,16 @@ const ChatRoomScreen: React.FC<ChatRoomScreenParams> = ({ history, chatId }) => 
               .toString(36)
               .substr(2, 9),
             createdAt: new Date(),
+            chat: {
+              __typename: 'Chat',
+              id: chatId
+            },
             content
           }
         },
         // tslint:disable-next-line:no-shadowed-variable
         update: (client: any, { data: { addMessage } }: any) => {
-          interface FullChat { [key: string]: any; }
-          let fullChat;
-
-          const chatIdFromStore = defaultDataIdFromObject(chat);
-
-          if (chatIdFromStore === null) {
-            return;
-          }
-
-          try {
-            fullChat = client.readFragment({
-              id: chatIdFromStore,
-              fragment: fragments.fullChat,
-              fragmentName: 'FullChat'
-            });
-          } catch (e) {
-            return;
-          }
-
-          if (fullChat === null || fullChat.messages === null) {
-            return;
-          }
-          if (fullChat.messages.some((m: any) => m.id === addMessage.id)) { return; }
-
-          fullChat.messages.push(addMessage);
-          fullChat.lastMessage = addMessage;
-
-          client.writeFragment({
-            id: chatIdFromStore,
-            fragment: fragments.fullChat,
-            fragmentName: 'FullChat',
-            data: fullChat
-          });
-
-          // tslint:disable-next-line:no-shadowed-variable
-          let data: ChatsQuery | null;
-
-          try {
-            data = client.readQuery({
-              query: queries.chats
-            });
-          } catch (e) {
-            return;
-          }
-
-          if (!data || !data.chats) {
-            return null;
-          }
-
-          const chats = data.chats;
-
-          const chatIndex = chats.findIndex((c: any) => {
-            if (addMessage === null || addMessage.chat === null) { return -1; }
-            return c.id === addMessage.chat.id;
-          });
-
-          if (chatIndex === -1) {
-            return;
-          }
-
-          const chatWhereAdded = chats[chatIndex];
-
-          chatWhereAdded.lastMessage = addMessage;
-          // The chat will appear at the top of the ChatsList component
-          chats.splice(chatIndex, 1);
-          chats.unshift(chatWhereAdded);
-
-          client.writeQuery({
-            query: queries.chats,
-            data: { chats }
-          });
+          writeMessage(client, addMessage);
         }
       });
     },
