+++ client/src/graphql/fragments/message.fragment.ts
@@ -4,5 +4,8 @@ export default gql`
     id
     createdAt
     content
+    chat {
+      id
+    }
   }
 `;
