+++ client/src/services/cache.service.ts
@@ -1,0 +1,92 @@
+import { DataProxy } from 'apollo-cache';
+import { defaultDataIdFromObject } from 'apollo-cache-inmemory';
+import { ApolloClient } from 'apollo-client';
+import * as fragments from '../graphql/fragments';
+import * as queries from '../graphql/queries';
+import { MessageFragment, useMessageAddedSubscription } from '../graphql/types';
+
+type Client = ApolloClient<any> | DataProxy;
+
+export const useCacheService = () => {
+  useMessageAddedSubscription({
+    onSubscriptionData: ({ client, subscriptionData: { data } }) => {
+      if (data) {
+        writeMessage(client, data.messageAdded);
+      }
+    }
+  });
+};
+
+export const writeMessage = (client: Client, message: MessageFragment) => {
+  type FullChat = { [key: string]: any; };
+  let fullChat;
+
+  const chatIdFromStore = defaultDataIdFromObject(message.chat);
+
+  if (chatIdFromStore === null) {
+    return;
+  }
+
+  try {
+    fullChat = client.readFragment<FullChat>({
+      id: chatIdFromStore,
+      fragment: fragments.fullChat,
+      fragmentName: 'FullChat'
+    });
+  } catch (e) {
+    return;
+  }
+
+  if (fullChat === null || fullChat.messages === null) {
+    return;
+  }
+  if (fullChat.messages.some((m: any) => m.id === message.id)) { return; }
+
+  fullChat.messages.push(message);
+  fullChat.lastMessage = message;
+
+  client.writeFragment({
+    id: chatIdFromStore,
+    fragment: fragments.fullChat,
+    fragmentName: 'FullChat',
+    data: fullChat
+  });
+
+  // tslint:disable-next-line:no-shadowed-variable
+  let data;
+
+  try {
+    data = client.readQuery({
+      query: queries.chats
+    });
+  } catch (e) {
+    return;
+  }
+
+  if (!data || !data.chats) {
+    return null;
+  }
+
+  const chats = data.chats;
+
+  const chatIndex = chats.findIndex((c: any) => {
+    if (message === null || message.chat === null) { return -1; }
+    return c.id === message.chat.id;
+  });
+
+  if (chatIndex === -1) {
+    return;
+  }
+
+  const chatWhereAdded = chats[chatIndex];
+
+  chatWhereAdded.lastMessage = message;
+  // The chat will appear at the top of the ChatsList component
+  chats.splice(chatIndex, 1);
+  chats.unshift(chatWhereAdded);
+
+  client.writeQuery({
+    query: queries.chats,
+    data: { chats }
+  });
+};
